[SNC] Implement High-Rise Sawjack. Add BlocksCreatureWithFlyingTriggeredAbility.
