+++ Mage.Sets/src/mage/cards/f/FrostwebSpider.java
@@ -3,7 +3,7 @@ package mage.cards.f;
 
 import java.util.UUID;
 import mage.MageInt;
-import mage.abilities.TriggeredAbilityImpl;
+import mage.abilities.common.BlocksCreatureWithFlyingTriggeredAbility;
 import mage.abilities.common.delayed.AtTheEndOfCombatDelayedTriggeredAbility;
 import mage.abilities.effects.Effect;
 import mage.abilities.effects.common.CreateDelayedTriggeredAbilityEffect;
@@ -16,13 +16,8 @@ import mage.constants.CardType;
 import mage.constants.SubType;
 import mage.constants.SuperType;
 import mage.counters.CounterType;
-import mage.constants.Zone;
-import mage.filter.FilterPermanent;
 import mage.filter.common.FilterCreaturePermanent;
 import mage.filter.predicate.mageobject.AbilityPredicate;
-import mage.game.Game;
-import mage.game.events.GameEvent;
-import mage.game.permanent.Permanent;
 /**
  *
  * @author L_J
@@ -48,7 +43,7 @@ public final class FrostwebSpider extends CardImpl {
         // Whenever Frostweb Spider blocks a creature with flying, put a +1/+1 counter on Frostweb Spider at end of combat.
         Effect effect = new CreateDelayedTriggeredAbilityEffect(new AtTheEndOfCombatDelayedTriggeredAbility(new AddCountersSourceEffect(CounterType.P1P1.createInstance())), true);
         effect.setText("put a +1/+1 counter on {this} at end of combat");
-        this.addAbility(new FrostwebSpiderTriggeredAbility(effect, filter, false));
+        this.addAbility(new BlocksCreatureWithFlyingTriggeredAbility(effect, false));
     }
 
     private FrostwebSpider(final FrostwebSpider card) {
@@ -58,46 +53,5 @@ public final class FrostwebSpider extends CardImpl {
     @Override
     public FrostwebSpider copy() {
         return new FrostwebSpider(this);
-    }
-}
-
-class FrostwebSpiderTriggeredAbility extends TriggeredAbilityImpl {
-
-    protected FilterPermanent filter;
-
-    public FrostwebSpiderTriggeredAbility(Effect effect, FilterPermanent filter, boolean optional) {
-        super(Zone.BATTLEFIELD, effect, optional);
-        this.filter = filter;
-    }
-
-    public FrostwebSpiderTriggeredAbility(final FrostwebSpiderTriggeredAbility ability) {
-        super(ability);
-        this.filter = ability.filter;
-    }
-
-    @Override
-    public boolean checkEventType(GameEvent event, Game game) {
-        return event.getType() == GameEvent.EventType.BLOCKER_DECLARED;
-    }
-
-    @Override
-    public boolean checkTrigger(GameEvent event, Game game) {
-        if (event.getSourceId().equals(this.getSourceId())) {
-            Permanent blocked = game.getPermanent(event.getTargetId());
-            if (blocked != null && filter.match(blocked, game)) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    @Override
-    public String getTriggerPhrase() {
-        return "Whenever {this} blocks a " + filter.getMessage() + ", " ;
-    }
-
-    @Override
-    public FrostwebSpiderTriggeredAbility copy() {
-        return new FrostwebSpiderTriggeredAbility(this);
     }
 }
