+++ Mage.Sets/src/mage/cards/s/Snarespinner.java
@@ -1,19 +1,14 @@
 package mage.cards.s;
 
 import mage.MageInt;
-import mage.abilities.TriggeredAbilityImpl;
+import mage.abilities.common.BlocksCreatureWithFlyingTriggeredAbility;
 import mage.abilities.effects.common.continuous.BoostSourceEffect;
-import mage.abilities.keyword.FlyingAbility;
 import mage.abilities.keyword.ReachAbility;
 import mage.cards.CardImpl;
 import mage.cards.CardSetInfo;
 import mage.constants.CardType;
 import mage.constants.Duration;
 import mage.constants.SubType;
-import mage.constants.Zone;
-import mage.game.Game;
-import mage.game.events.GameEvent;
-
 import java.util.UUID;
 
 /**
@@ -32,7 +27,7 @@ public final class Snarespinner extends CardImpl {
         this.addAbility(ReachAbility.getInstance());
 
         // Whenever Snarespinner blocks a creature with flying, Snarespinner gets +2/+0 until end of turn.
-        this.addAbility(new SnarespinnerTriggeredAbility());
+        this.addAbility(new BlocksCreatureWithFlyingTriggeredAbility(new BoostSourceEffect(2, 0, Duration.EndOfTurn), false));
     }
 
     private Snarespinner(final Snarespinner card) {
@@ -42,37 +37,5 @@ public final class Snarespinner extends CardImpl {
     @Override
     public Snarespinner copy() {
         return new Snarespinner(this);
-    }
-}
-
-class SnarespinnerTriggeredAbility extends TriggeredAbilityImpl {
-
-    SnarespinnerTriggeredAbility() {
-        super(Zone.BATTLEFIELD, new BoostSourceEffect(2, 0, Duration.EndOfTurn), false);
-    }
-
-    private SnarespinnerTriggeredAbility(final SnarespinnerTriggeredAbility ability) {
-        super(ability);
-    }
-
-    @Override
-    public boolean checkEventType(GameEvent event, Game game) {
-        return event.getType() == GameEvent.EventType.BLOCKER_DECLARED;
-    }
-
-    @Override
-    public boolean checkTrigger(GameEvent event, Game game) {
-        return event.getSourceId().equals(this.getSourceId())
-                && game.getPermanent(event.getTargetId()).getAbilities().containsKey(FlyingAbility.getInstance().getId());
-    }
-
-    @Override
-    public String getRule() {
-        return "Whenever {this} blocks a creature with flying, {this} gets +2/+0 until end of turn.";
-    }
-
-    @Override
-    public SnarespinnerTriggeredAbility copy() {
-        return new SnarespinnerTriggeredAbility(this);
     }
 }
